Move scrollLogRegistry to shared/utils to fix cross-panel import
features/sessions/useSessionLifecycle was importing scrollLogRegistry
directly from panels/agent/utils — a cross-layer violation. Extract
the registry to shared/utils/scrollLogRegistry.ts and re-export from
the agent panel's scrollLog.ts so existing agent-internal imports
still work.
