+++ src/renderer/features/sessions/hooks/useSessionLifecycle.ts
@@ -5,7 +5,7 @@ import { useEffect, useCallback, useState } from 'react'
 import type { Session } from '../../../store/sessions'
 import type { ProfileData } from '../../../store/profiles'
 import { terminalBufferRegistry } from '../../../shared/utils/terminalBufferRegistry'
-import { scrollLogRegistry } from '../../../panels/agent/utils/scrollLog'
+import { scrollLogRegistry } from '../../../shared/utils/scrollLogRegistry'
 import { loadMonacoProjectContext } from '../../../shared/utils/monacoProjectContext'
 import { restoreSessionFocus } from '../../../shared/utils/focusHelpers'
 import { fetchReviewStatus } from '../../../shared/utils/reviewStatus'
